transactionStore goToAdd, goToTransaction, goToDetail 추가
