+++ src/components/RecentHistory.vue
@@ -14,23 +14,18 @@
         :index="index"
       />
     </ul>
-    <button @click="goToAdd">추가하기</button>
-    <button @click="goToTransaction">더보기</button>
+    <button @click="store_trans.goToAdd">추가하기</button>
+    <button @click="store_trans.goToTransaction">더보기</button>
   </div>
 </template>
 
 <script setup>
-import { useRouter } from 'vue-router';
-import { useCategoryStore } from '@/stores/categoryStore';
 import { useTransactionStore } from '@/stores/transactionStore';
 import { storeToRefs } from 'pinia';
 import { ref, computed } from 'vue';
 import TransactionItem from './TransactionItem.vue';
 
-const store_category = useCategoryStore();
 const store_trans = useTransactionStore();
-
-const router = useRouter();
 
 store_trans.fetchTransactions();
 const { transactions } = storeToRefs(store_trans);
@@ -48,13 +43,6 @@ const recentTransactions = computed(() => {
 
   return filtered.slice(0, 5);
 });
-
-const goToAdd = () => {
-  router.push('/transaction/add');
-};
-const goToTransaction = () => {
-  router.push('/transaction');
-};
 </script>
 
 <style scoped></style>
