+++ src/components/TransactionItem.vue
@@ -1,5 +1,8 @@
 <template>
-  <li @click="goToDetail" style="list-style: none; cursor: pointer">
+  <li
+    @click="store_trans.goToDetail(trans.id)"
+    style="list-style: none; cursor: pointer"
+  >
     <!-- 클릭하면 상세 내역 보기 -->
     <span>
       {{ index + 1 }} {{ trans.date }} {{ trans.amount }} {{ trans.category }}
@@ -9,18 +12,14 @@
 </template>
 
 <script setup>
-import { useRouter } from 'vue-router';
+import { useTransactionStore } from '@/stores/transactionStore';
+
+const store_trans = useTransactionStore();
 
 const props = defineProps({
   trans: Object,
   index: Number,
 });
-
-const router = useRouter();
-
-const goToDetail = () => {
-  router.push(`/transaction/${props.trans.id}`);
-};
 </script>
 
 <style scoped></style>
